Moved checkAuth into App
